AGNT-9: Use master branch only (#152)
To decouple SBE/Agent releases, start adding information specific
information in the specification.

x-since is not used by any tool but is here as a meta information that
could potentially be used in the future. Therefore we also set this as
part of the description.

Backport this approach to changes from 20.14 version.