AGNT-9: Drop snapshot in Swagger spec version
As we are using master branch only, just drop it and use the next
version of the Agent instead. Once we start putting new features in the
spec we can increment this number manually too.

Upon a release when we synchronize the internal specs with those being
public we will put anyway the released version of the Agent or SBE.