Migrating agent APIs to OpenApi 3 (#158)
Goal of this commit is to migrate Agent APIs to start using OpenAPI 3 instead of swagger 2.
Swagger 2 version is still available to maintain compatibility with exisiting systems that are using the specs to generate code but it will not be updated with the new coming features. Only OpenApi 3 version will be kept up to date with new implementations from now on.